+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @khanacademy/wonder-blocks-popover
 
+## 1.1.25
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-icon-button@3.4.16
+-   @khanacademy/wonder-blocks-modal@3.0.2
+-   @khanacademy/wonder-blocks-tooltip@1.4.1
+
+## 1.1.24
+
+### Patch Changes
+
+-   Updated dependencies [b784d7a8]
+    -   @khanacademy/wonder-blocks-tooltip@1.4.0
+
 ## 1.1.23
 
 ### Patch Changes